Added support for zookeeper "id"
The order of zookeeper nodes listed in the hosts file
is not preserved when generating the server.id entries
in the zookeeper.properties files.
This is because Ansible uses a standard Python dict, which
is hash based and does not preserve insertion order.

This is normally not an issue except when we are trying to define
ZooKeeper groups, which relies on the ids.

This change makes it possible to add and use an artificial id to
each ZooKeeper entry

zookeeper:
hosts:
zookeeper-01:
zookeeper:
id: 1
zookeeper-02:
zookeeper:
id: 2

If the id is not provided, the template will fall back to using
the loop index to assign an id.

This closes #77